Relocating Version Check and specified default Version for now
